replace getEvtWeight with cacheStatic* + computeNonStatic*
Details:

getEvtWeight was a mix of factors that depend only on the event, and
factors that depend on the selected objects.  It is now split in two
parts:
+ cacheStaticWeightComponents
+ computeNonStaticWeightComponents
In particular, the cacheStaticWeightComponents will cache the value
coming from SusyNtAna::getEventWeight, whose call seems to be fairly
expensive (map lookup? we might want to cache it upstream...).