reset weights within SusySelection::cacheStaticWeightComponents
Details:
There is already a call to reset the weight components within Process.
That is useful only within SusySelection, and it is necessary for the
increment[readin].  However, this doesn't work for the inheriting
classes (e.g. SusyPlotter). This call within
cacheStaticWeightComponents should ensure that we start with unity
weight components at each event.