move the enums within SusyPlotter
Details:

these enums (and the corresponding arrays of strings), conflict with Matt's
definitions. Move them within the SusyPlotter scope, to avoid the conflict.
One thing that I really don't like is that these should be static
datamember, but rootcint cannot deal with them (see
http://root.cern.ch/phpBB3/viewtopic.php?t=8819).
We could also do a better initialization of the arrays (rather than an
assignement as doing now). I've done this in the past, but I cannot
remember the trick (it was something like an intermediate vector, then
array c'tor from vector). Fine for now.