apply d0sig and muiso cuts upstream in isSignalMuon + isSignalElectron
Details:
We are currently using this special criteria to select our signal
leptons.  Because the fake estimate heavily relies on the m_signal*
collections, Matt suggested that I apply these selection criteria
upstream, within SusyNtuple.  This required the definition of a new
element in the AnaType enum, and a few extra switches in SusyNtTools.
The modifications are in the file doc/SusyNtuple.patch2013-09-11, and
they should be applied on top of SusyNtuple-00-01-04.  If we see that
these two criteria are really providing an improvement in our
sensitivity, then we should commit it to SusyNtTools.

I tested (with MeasureFakeRate2) that I see the changes we expect
(i.e. same yields, but now the muoniso and the eld0 requirement do not
have any effect).