ugly hack to avoid SR/CR labeling mess. to be removed ASAP.
Details:
I am really confused by the names of the CR/SR, where some of the CR
are labeled as 'SR*', and some of the SR are labeled as 'CR*'. The
bottom line is that here the code is trying to access
'muon_SR_WHSS_all_flavor_den' and the file contains
'muon_CR_WHSS_all_flavor_den'.
Ask Matt about it.