FakePlotting::DataMCSF : update input file
Details:
Use my input file. This function, however, is still crashing.
It seems to crash on 'h_names.push_back(crname);', and I don't have
any idea about why it should do this...  Matt says that this function
is mostly for debugging and cross-check, and that it might not be up
to date.  Putting it on hold, for now.