refactor selectEvent ad passSrSs
Details:
Refactor SusySelection::selectEvent and SusySelection::passSrSs,
splitting each of these functions in two logical steps:
- determine the flags
- increment the counters
This split is necessary when looping over systematics, so that
we don't increment N times the same counters.
This is a large commit, so here are the technical details:
- split selectEvent in computeEventFlags, incrementCounters
- split passSrSs in computeSsFlags, incrementCounters
- add the struct EventFlags

Tested with SusyPlotter that the cutflow is unchanged (but
modified other classes as well).