refactor HftFiller::fill and define WeightVariations
- define one fill() method only, and make the new
  HftFiller::assignDilepVars private.
- define a new WeightVariations struct

In this way we can hide the details of how we assign the kinematic
variables and the run/event (both always present), and reuse this code
in two public `fill` methods, one with and one without
`WeightVariations`. This allows to control the default values of the
weight variations when they are not computed (i.e. we always assign
them).