add a script to map the dsid to the group; add also the generated header DsidGroups.h
This script is needed because we are specifying the xsec uncertainties
for each group. However, when we run we only have access to the dsid,
so we need to create a map. This script creates a header definig such
map. The groups and the xsec uncertainties are are defined separately.