add protection against unphysical qflip probabilities
We have a handful of events (see for example dsid 177997) for which
the chargeFlip tool returns a negative probability (-1e-5) and a
handful of them for which it returns a very large probability (1e5).
This should be fixed upstram in the tool; however, since it breaks
HistFitter, I am adding a temporary protection here.