add the mc generator uncertainties
These uncertainties are now stored in the trees (needs the branches
GENUP, GENDOWN). The WZ uncertainty is the one determined by Steve,
for the other samples we just use a flat 100%.
Using the 'groupFromDsid' implemented for the xsec uncertainty to
determine whether this is WZ or something else.
Note that 100% is also the default value.