SusyPlotter, MatrixPrediction: do not fill old histos
If I book all histograms for all the old regions when I also create
the hft trees with all systematics, I run out of memory. So I decided
to temporarily disable the old histograms (and selection regions).
Later on I will remove this part of the code (and do everything from
the trees).