switch to MCWeighter; switch to SusyNtTools from github
I need MCWeighter to make the pmssm hft trees. The issue with the
sumw=0 normalization is now sorted out, but it's available from github
and in this area I am still using the svn repo, and the old
normalization. Now switching to github, and also pulling in the jvf
uncertaintyes. Details:

- define `initMcWeighter()`
- change the signature of all the following functions, passing in `jvfTool` and `sys`:
    - computeSsFlags, filterClJets, isCentralLightJet, numberOfCLJets, assignNjetFlags
- drop obsolete passEwkSs, passEwkSsLoose, passEwkSsLea

This code compiles on top of `SusyNtuple-00-01-11-12-g4f2ce50`, that is
Merge remote-tracking branch 'origin/dbg-mcweighter' into dbg-mcweighter