MeasureFakeRate2::passMCReg: switch back to metrel>40 extraction region
At some point (616d064) for the same-sign study I switched to a
different extraction region (with >=1j instead of a metrel cut). Now
for the razor we probably want to go back to the original extraction
region (and perhaps also for the LFV). Need to check what the fake
scale factors are going to be with this selection.